fix : make map compo
